% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add.points}
\alias{add.points}
\title{Manually specify the value of a point in the skeleton}
\usage{
add.points(bm, t, W_t, label = names(t))
}
\arguments{
\item{t}{time}

\item{W_t}{value}
}
\description{
Only valid outside layers. Cannot be existing time (should delete then add)
}
