% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_localise.R
\name{bb.localise}
\alias{bb.localise}
\title{Create localised layer information around a Brownian bridge}
\usage{
bb.localise(bm, s, t, refine = bm$refine, mult = bm$mult, prefer = bm$prefer)
}
\arguments{
\item{s}{start time}

\item{t}{end time

The function will create localised layers across all currently layer-free Brownian bridges between times s and t}
}
\description{
Create localised layer information around a Brownian bridge
}
