% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_layers.R
\name{bessel.layers}
\alias{bessel.layers}
\title{Construct Bessel layers}
\usage{
bessel.layers(bm, s, t)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{s}{left hand time point}

\item{t}{right hand time point}
}
\value{
The Brownian motion object which was passed in argument \code{bm} is
updated in place and returned, enabling chaining of commands with
dplyr (and other) style pipes.
}
\description{
Unbiasedly simulates the smallest Bessel layers which contains the Brownian
motion between two known sample points by retrospective Bernoulli sampling
and inversion sampling.
}
