% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bm.R
\name{create.bm}
\alias{create.bm}
\title{Create a Brownian motion object}
\usage{
create.bm(t = 0, W_t = 0, refine = TRUE, mult = 1, prefer = "bessel")
}
\arguments{
\item{t}{vector of fixed times of the Brownian motion.  Defaults to 0.}

\item{W_t}{vector of corresponding locations of the Brownian motion at the times t.  Defaults to 0.}

\item{refine}{refine is}

\item{mult}{mult is}

\item{prefer}{prefer is}
}
\description{
Creates an R environment to contain the trajectories and layer information of
a Brownian motion.
}
