% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{delete.labels}
\alias{delete.labels}
\title{Manually delete labels after simulation}
\usage{
delete.labels(bm, t = NA, label = NA)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place}

\item{t}{vector of fixed times of the Brownian motion.}

\item{label}{vector of length 1 or the length of t indicating the introduction of user specified labels (if any). By default, uses any names in the vector of times.}
}
\description{
Description of deleting labels
}
