% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_skeleton.R
\name{delete.skeleton}
\alias{delete.skeleton}
\title{Eliminate parts of path, layer or both from the skeleton}
\usage{
delete.skeleton(bm, l = -Inf, r = Inf, type = "all")
}
\arguments{
\item{l}{left end of open interval to delete.  By default, `-Inf` which
results in left truncation of the path up to `r`}

\item{r}{right end of open interval to delete.  By default, `Inf` which
results in right truncation of the path from `l`}

\item{type}{one of `"all"` or `"layer"` to specify whether to delete both path
observations and layers, or just layers in the open interval `(l,r)`}
}
\description{
Eliminate parts of path, layer or both from the skeleton
}
