% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_passage.R
\name{first.passage}
\alias{first.passage}
\title{Find a first passage time and state}
\usage{
first.passage(
  bm,
  l = NULL,
  u = NULL,
  delta.l = NULL,
  delta.u = NULL,
  delta = NULL,
  label = NULL
)
}
\arguments{
\item{bm}{a Brownian motion object (of dimension 1) from which simulation should continue.
Note the object is updated in place.}

\item{l}{scalar giving lower bound for first passage.}

\item{u}{scalar giving upper bound for first passage.}

\item{delta.l}{provide bounds via the negative change in Brownian motion. May be specified together with \code{delta.u} to provide asymmetric first passage delta.}

\item{delta.u}{provide bounds via the positive change in Brownian motion. May be specified together with \code{delta.l} to provide asymmetric first passage delta.}

\item{delta}{scalar providing the bounds by way of the positive/negative change in the Brownian motion.}

\item{label}{a user specified label for the first passage time created by this layer (optional).}
}
\value{
The Brownian motion object which was passed in argument \code{bm} is
updated in place and returned, enabling chaining of commands with
dplyr (and other) style pipes.
}
\description{
Finds the first passage time of a Brownian motion object to a specified lower,
upper, or lower and upper limit.
}
\details{
This function updates a Brownian motion object adding the first passage time to
the specified lower (\code{l}), upper (\code{u}), or lower and upper limits.
The limits can also be specfied via \code{delta} which places the limits as the
current endpoint state +/- \code{delta}.
Note that simulation of the first passage implicitly imposes constraints on the
underlying Brownian motion between the endpoint and the first passage
time.

In particular, in the case that the lower and upper limits are asymmetric around
the endpoint, there may be a recursive simulation of nested symmetric first passages
between the endpoint and the ultimate first passage, resulting in implicit imposition
of nested constraints.  When plotting a Brownian motion these are visible as dashed
red lines nested within the solid red line constraint imposed by the overall first
passage simulation.
}
