% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_layers.R
\name{intersection.layers}
\alias{intersection.layers}
\title{Simulate an intersection layer}
\usage{
intersection.layers(
  bm,
  s,
  t,
  refine = bm$refine,
  mult = bm$mult,
  prefer = bm$prefer
)
}
\description{
Simulate an intersection layer
}
