% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Construct new layers}
\usage{
layers(
  bm,
  s,
  t,
  type = bm$prefer,
  refine = bm$refine,
  mult = bm$mult,
  prefer = bm$prefer,
  label = c(names(s), names(t))
)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place}

\item{s}{left hand time point}

\item{t}{right hand time point}

\item{type}{type specifies whether to carry out "bessel" or "intersection" or "localised" simulation. If of length 1, then this specifies the type for every dimension. If of length dim, then this specifies for each dimension the type. By default this matches the setting for preference specified on the creation of the Brownian motion object.}

\item{refine}{indicates whether refinement should be by default carried out. If of length 1, then this specifies refinement for every dimension. If of length dim, then this specifies refinement for each dimension (in order). By default uses the setting on creation of the Brownian motion object.}

\item{mult}{the default layer size is \code{sqrt(t-s)}.  You can scale
this by specifying the \code{mult} argument which will result in layer
sizes of \code{mult*sqrt(t-s)}. By default uses the setting on creation of the Brownian motion object.}

\item{prefer}{indicates whether there is a preference for "bessel" or "intersection" layers where possible. If of length 1, then this specifies the preference for every dimension. If of length dim, then this specifies preference for each dimension (in order). By default uses the setting on creation of the Brownian motion object.}

\item{label}{vector of length 1 or the length of t indicating the introduction of user specified labels (if any). By default, uses any names in the vector of times.}
}
\value{
The Brownian motion object which was passed in argument \code{bm} is
updated in place and returned, enabling chaining of commands with
dplyr (and other) style pipes.
}
\description{
TODO UPDATE ME
Performs unbiased simulation of the smallest specified layers which contain the
Brownian motion between two known sample points by retrospective Bernoulli
sampling and inversion sampling.
}
