% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{plotBM}
\alias{plotBM}
\title{Plot Brownian motion simulation}
\usage{
plotBM(x, type = "l", pointcolor = c("purple", "blue"), cex = 2, ...)
}
\arguments{
\item{x}{A vector or a matrix of Brownian motion path}

\item{type}{Plotting type (line/points/etcs) which is passed to plot() function}

\item{pointcolor}{Colors of starting and ending points}

\item{cex}{Points size for the starting and ending points}

\item{...}{Other parameters passed to plot() function}
}
\value{
Brownina motion simulation path figure
}
\description{
Plot 1-dimension or 2-dimension Brownian motion simulation path
}
\details{
Plot 1-dimension or 2-dimension Brownian motion simulation path
}
\examples{
x <- sim1dBM(1000, 2)
plotBM(x)
}
\author{
Ye Zheng
}

