% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine.R
\name{refine}
\alias{refine}
\title{Refine an intersection layer}
\usage{
refine(bm, s, t, mult = bm$mult)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place}

\item{s}{left hand time point}

\item{t}{right hand time point}

\item{mult}{the default layer size is \code{sqrt(t-s)}.  You can scale
this by specifying the \code{mult} argument which will result in layer
sizes of \code{mult*sqrt(t-s)}. By default uses the setting on creation of the Brownian motion object.}
}
\description{
Refine an intersection layer
}
