% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_points.R
\name{set.points}
\alias{set.points}
\title{Manually specify the value of a point in the skeleton}
\usage{
set.points(bm, t, W_t, add.segment = FALSE, label = names(t))
}
\arguments{
\item{t}{time}

\item{W_t}{value}

\item{add.segment}{TRUE/FALSE.  If true, then this creates a jump
discontinuity at the specified time by updating only W_t and not W_tm.
If false, then this updates both W_t and W_tm.
Ignored if t is a new rather than existing time in the skeleton.}
}
\description{
Manually specify the value of a point in the skeleton
}
