% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bb.R
\name{sim.bb}
\alias{sim.bb}
\title{Simulate a Brownian bridge}
\usage{
sim.bb(bm, s, t, q = NULL, q.grid = NULL, label = names(q))
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{s}{a scalar time upon which to condition the left hand end of the
Brownian bridge.  This time must have already been simulated and be present
in the Brownian motion object supplied in the \code{bm} argument.}

\item{t}{a scalar time upon which to condition the right hand end of the
Brownian bridge.  This time must have already been simulated and be present
in the Brownian motion object supplied in the \code{bm} argument.}

\item{q}{a vector of times at which to simulate the Brownian bridge, or NULL
(the default) if the argument \code{q.grid} is to be used.  Note that all
times must be strictly contained in the interval from \code{s} to \code{t}.}

\item{q.grid}{how many times to simulate the Brownian bridge at, which will
be equally spaced between \code{s} and \code{t}, or NULL (the default) if
the argument \code{q} is to be supplied.}
}
\value{
the Brownian motion object which was passed in argument \code{bm} is
  updated in place and returned, enabling chaining of commands with
  dplyr (and other) style pipes.
}
\description{
Simulates Brownian motion conditional upon two (previously simulated) endpoints, outside of any layer.
Exactly one of the arguments \code{q} or \code{q.grid} should be supplied.
}
