% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cond_bessel.R
\name{sim.condbessel}
\alias{sim.condbessel}
\title{Simulate Brownian motion conditional on bessel layer}
\usage{
sim.condbessel(bm, s, t, q = NULL, q.grid = NULL)
}
\description{
This currently converts the Bessel layer to an intersection layer and then
uses the conditional intersection layer sampling.
}
