% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cond_intersection.R
\name{sim.condintersection}
\alias{sim.condintersection}
\title{Simulate Brownian motion conditional on intersection layer}
\usage{
sim.condintersection(bm, s, t, q = NULL, q.grid = NULL, label = names(q))
}
\description{
Simulate Brownian motion conditional on intersection layer
}
