% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cond_local.R
\name{sim.condlocal}
\alias{sim.condlocal}
\title{Simulate Brownian motion conditional on localised layer}
\usage{
sim.condlocal(bm, s, t, q = NULL, q.grid = NULL)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}
}
\value{
the Brownian motion object which was passed in argument \code{bm} is
  updated in place and returned, enabling chaining of commands with
  dplyr (and other) style pipes.
}
\description{
Simulates Brownian motion
}
