% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate Brownian motion at a target time}
\usage{
sim(bm, t, refine = bm$refine, mult = bm$mult, prefer = bm$prefer)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{t}{a vector of times to simulate at.}

\item{refine}{whether to automatically refine layers where the simulated
time results in bisection of an existing layer.  Defaults to the option
specified at the time of creation of \code{bm}.}

\item{mult}{the mult parameter to be passed through to any layer operations
including for the level of refinement if \code{refine=TRUE}.   Defaults to the option
specified at the time of creation of \code{bm}.}

\item{prefer}{indicate the preferred layer type which should arise after
conditional simulation bisects an existing layer (this will be respected
where possible, but it not always achievable).    Defaults to the option
specified at the time of creation of \code{bm}.}
}
\value{
the Brownian motion object which was passed in argument \code{bm} is
  updated in place and returned, enabling chaining of commands with
  dplyr (and other) style pipes.
}
\description{
Simulates Brownian motion at chosen times, automatically applying the correct
algorithm to condition on the current known (ie previously simulated) states
of the trajectory.
}
