% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_uncond.R
\name{sim.uncond}
\alias{sim.uncond}
\title{Simulate Brownian motion unconditionally}
\usage{
sim.uncond(bm, t, t.grid = NULL)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{t}{either a vector of times to simulate at, or a single endpoint time
if t.grid is supplied.}

\item{t.grid}{by default this is NULL so that only the times supplied in
\code{t} are simulated.  If a scalar integer is supplied then an evenly
spaced grid of this many times will be simulated up to \code{t}.}
}
\value{
the Brownian motion object which was passed in argument \code{bm} is
  updated in place and returned, enabling chaining of commands with
  dplyr (and other) style pipes.
}
\description{
Simulates Brownian motion unconditionally at times beyond the current
endpoint stored in the supplied object.  Providing times in \code{t} which
are not beyond the latest simulated endpoint will result in an error.
}
