% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim1dBM}
\alias{sim1dBM}
\title{1-dimension Brownian motion simulation}
\usage{
sim1dBM(n, sigma = 1)
}
\arguments{
\item{n}{Number of points to simulate(must be larger than 1 point)}

\item{sigma}{The standard deviation of the step}
}
\value{
A vector showing the Brownian motion
}
\description{
Simulate 1-dimension Brownian motion
}
\details{
Simulate 1-dimension Brownian motion
}
\examples{
x <- sim1dBM(1000, 2)
}
\author{
Ye Zheng
}

