% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim2dBM}
\alias{sim2dBM}
\title{2-dimension Brownian motion simulation}
\usage{
sim2dBM(n, sigma = 1)
}
\arguments{
\item{n}{Number of points to simulate(must be larger than 1 point)}

\item{sigma}{The standard deviation of the step}
}
\value{
Matrix with two columns showing the (X, Y) positions of simulated Brownina modtion
}
\description{
Simulate 2-dimension Brownian motion
}
\details{
Simulate 2-dimension Brownian motion
}
\examples{
x <- sim2dBM(1000, 2)
}
\author{
Ye Zheng
}

