% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\title{Return process state}
\usage{
state(bm, t = NULL)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{t}{a scalar time at which the state is required.  This time must have
already been simulated and be present in the Brownian motion object
supplied in the \code{bm} argument.  If omitted, the most recent state is
returned.}
}
\value{
a vector containing the state at the specified time.
}
\description{
Returns the state of the Brownian motion at a given time.
}
