% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_segment.R
\name{add.segment}
\alias{add.segment}
\title{Create/move a segment}
\usage{
add.segment(bm, l = -Inf, r = Inf, delta = NULL, W_t = NULL, label = NULL)
}
\arguments{
\item{l}{}
}
\description{
Create/move a segment
}
