% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_bm.R
\name{concat.bm}
\alias{concat.bm}
\title{Copy / Concatenate Brownian motion objects}
\usage{
concat.bm(..., t0 = NULL)
}
\arguments{
\item{...}{multiple Brownian motion objects which will be concatenated together in the order provided into a new BrownianMotion object.
Note the object is updated in place}

\item{t0}{scalar representing the time the new Brownian motion should be initialised.  Defaults to value of initial time of the first Brownian motion object to be concatenated.}
}
\description{
Copy / Concatenate Brownian motion objects
}
