% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bm.R
\name{create.bm}
\alias{create.bm}
\title{Create a Brownian motion object}
\usage{
create.bm(W_0 = 0)
}
\arguments{
\item{W_0}{the starting location of the Brownian motion at time 0.}
}
\description{
Creates an R environment to contain the trajectories and layer information of
a Brownian motion.
}
