% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Construct new layers}
\usage{
layers(
  bm,
  s,
  t,
  type = bm$prefer,
  refine = bm$refine,
  mult = bm$mult,
  prefer = bm$prefer,
  label = c(names(s), names(t))
)
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place}

\item{s}{left hand time point}

\item{t}{right hand time point}

\item{mult}{the default layer size is \code{sqrt(t-s)}.  You can scale
this by specifying the \code{mult} argument which will result in layer
sizes of \code{mult*sqrt(t-s)}.}
}
\value{
The Brownian motion object which was passed in argument \code{bm} is
updated in place and returned, enabling chaining of commands with
dplyr (and other) style pipes.
}
\description{
TODO UPDATE ME
Performs unbiased simulation of the smallest specified layers which contain the
Brownian motion between two known sample points by retrospective Bernoulli
sampling and inversion sampling.
}
