% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_uncond.R
\name{sim.uncond}
\alias{sim.uncond}
\title{Simulate Brownian motion unconditionally}
\usage{
sim.uncond(bm, t, label = names(t))
}
\arguments{
\item{bm}{a Brownian motion object from which simulation should continue.
Note the object is updated in place.}

\item{t}{a vector of times at which to simulate the path}
}
\value{
the Brownian motion object which was passed in argument \code{bm} is
  updated in place and returned, enabling chaining of commands with
  dplyr (and other) style pipes.
}
\description{
Simulates Brownian motion unconditionally at times beyond the current
endpoint stored in the supplied object.  Providing times in \code{t} which
are not beyond the latest simulated endpoint will result in an error.
}
